rutabaga_gfx: Support 2D scanout in cross-domain component
Implement 2D scanout in the CrossDomain component by leveraging RutabagaResource::new_2d and 2D transfer helpers in resource.rs.

Bug: 539486562
